requirements/base.txt
Django==2.2.25
django-cors-headers==3.10.1
django_model_utils==4.2.0
openpyxl==3.0.9
pandas==1.3.5
